requirements.txt
appnope==0.1.3
asgiref==3.5.2
backcall==0.2.0
decorator==5.1.1
Django==3.2.16
djangorestframework==3.14.0
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.2
matplotlib-inline==0.1.6
paho-mqtt==1.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
ptyprocess==0.7.0
py==1.11.0
Pygments==2.13.0
pytz==2022.6
retry==0.9.2
sqlparse==0.4.3
toml==0.10.2
traitlets==5.6.0
typing_extensions==4.4.0
wcwidth==0.2.5
